requirements.txt
Flask
pandas
opencv-python-headless
matplotlib
seaborn
torch
pillow
requests
pyyaml
tqdm
torchvision